requirement.txt
flask
gunicorn
tensorflow
numpy 
pandas
Pillow